requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.0
apscheduler>=3.10.0
imap-tools>=1.5.0
cryptography
